requirements.txt
requests>=2.25.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=0.19.0
retry>=0.9.2 